Option to use redis coordinator to find sequencer
This PR adds two new options for the autonomous-auctioneer which allow
it to discover the active sequencer so that it can send its auction
resolution transactions directly to it. If the sequencer is using the
redis coordinator to manage which sequencer is active, then the same
redis url can be passed as an option to the auctioneer.

The new options are:
--auctioneer-server.redis-coordinator-url
--auctioneer-server.use-redis-coordinator