Use the correct error
When the state provider isn't able to do its job because the chain is lagging
behind head, it's supposed to use a predefined error from the l2stateprovider
package. Instead, the nitro implemention of the provider was using a
redefinition of the error.

While this coding error easily accounts for why there were a bunch of ERROR logs
about the chain's not keeping up with the latest changes, it does not explain
why the test is sometimes flaky. I suspect that the original commits in this PR
were chasing the ERROR logs in the flaky tests, but that those were not the root
cause of the test failures.