Add a test for storage trie determinism
This includes a new BigMap contract in the nitro-contracts repo in the mocks
package.

The test attempts to set a lot of storage slots to some value in one
transaction, and then to clear 75% of them and add 10% more values in a second
transaction.

Finally, validation is run to ensure that the validator is behaving correctly.

The idea is that this test will pass validation with the state db's
deterministic flag set, but fail without it set.

Related to: NIT-3068