Hex encode bidder signatures in S3 published CSVs
Previously the signature field, which is just raw bytes, was being
written directly into the bid history CSVs which are published to S3 by
the auctioneer.

The uploaded files are for consumption by external parties and not used
by the auctioneer after they have been uploaded so no changes are needed
to read the files.

This change is being made pre-mainnet so we don't need to worry about
fixing existing files.