Use the eip4844 package instead of making rpc calls
The resoning goes like this. We need to be able to run arbitrum nodes against
public ethereum mainnet RPC endpoints, and those aren't likely to be running our
fork of go-ethereum, so they won't have these RPC methods exposed.

We have the information we need about the Ethereum Layer 1 chains against which
we may want to deploy and test things. For example, Sepolia, and Mainnet. We
also have a config we can use in tests with ChainID 1337.

Another alternative we discussed was making these configuration values
("MaxBlobGasPerBlock" and "BlobFeePerByte") part of the arbos state and allowing
ArbOwner to just set them explicitly. This would have the advantage of allowing
the ArbOwner to tweak the values to their liking, but it also introduces a way
for them to make incompatible configuraiton changes which may be harder to
debug.

We may still want to take this approach in the future. Especially, if we ever
decide to support blob posting on L2s. Currently, a nitro batch poster would
fail to post batches from a Layer 3 to a Layer 2 if it were configured
explicitly to post blobs. This is okay, for now, because we don't support
posting blobs to any L2.