Fix TestOpenInitializeChainDbEmptyInit
The problem is that upstream, pathdb stopped being permissive and silently
converting a root of `commit.Hash{}` to `types.EmptyRootHash` and our code
wasn't updated to explicitly initialize the arbitrum database with the expected
`types.EmptyRootHash`.

This commit also reverts a mistake I made on the previous commit on this branch
due to an incorrect warning in VSCode about not handling an error before
deferring some cleanup.