When not posting blobs, don't calculate blob posting costs
This change makes the batch_poster and data_poster stop looking up information
about the cost of posting blobs to the parent chain when they are configured not
to post blobs.

The way the batch_poster's Post4844Blobs config value is being passed down to
the data_poster isn't "smooth," but maybe it's good enough?