Restore batch_poster bounds for parent-chain EIP-7623 detection
This was temporarily changed when trying to update the test to also work with an
L3 that is running on an L2 with EIP-7623 enabled. But, there was something off
about the calculation when done in that scenario. We need to follow-up with a
proper understanding of how this should be modifed.

For now, the test was just removed, and I forgot to set the bounds back to 40
from 200.