Move the creation of the signer into the tx-processing loop
This is important because processing a transaction could result in a new arbos
version. Then, a later transaction in the same block, could start using the new
SetCodeTxType and that wouldn't be supported unless the code was also using the
updated Prague signer.

The reason this wasn't necessary before, is that the signer being used could be
derived from the header's time. And, in arbos 30-32, using the Cancun signer was
a no-op because arbitrum didn't support BlobTxType transactions.