Add a timer metric for tx inclusion
This change adds the following metrics:
arb_sequencer_tx_nontimeboosted_inclusion
arb_sequencer_tx_timeboosted_inclusion

They measure the time from a tx being published to the sequencer with
PublishTransaction or PublishTimeboostedTransaction, until when it is
successfully included in a block. In the Timeboosted case, this does
not include the time spent in the sequence number reordering buffer.

We don't want to record auction resolution transactions since they are
not technically Timeboosted but they bypass the delay, which would skew
the non-Timeboosted metric.

Note: The firstAppearance for a tx is moved forward to before the Timeboost
delay starts. The only thing that firstAppearance was used for
previously was how long the request could stay in the nonceFailureCache.
This effectively lowers that time by 200ms; if we think that's a problem
we can correct for it or use a different timestamp.