Fix the ensure-rollup-deployment flag causing the node to hang
The flag enables a feature which waits until the block where the rollup
was created is finalized. There is a bug right now which waits on the
new finalized blocks but the l1 reader is not initialized thus the
waiting hangs and the node does not start.

This is not present on when the node is started after the current finalized block
is ahead of the one which the node waits on. The reason is that the first check
does not wait on the channel of the l1reader for updates but fetches the current
finalized block instead thus it doesn't hit the uninialized reader bug.

The fix is to start the l1reader before subscribing.