Update to modern golangci-lint version
This updates the configuration file to work with versions of golangci-lint
greater than 2.0 and also upgrades the CI action to version 8 (from version 3).

For now, 8 of the static checks are disabled because enabling them makes large
diffs in our codebase. We can open a separate Linear ticket if we want to enable
some of those checks. (And, frankly, I would like that very much.)

Fixes NIT-3286