Limit parallel jobs within make to 8
Currently, we very frequently see the build step in the CI workflow run into:
`error: Resource temporarily unavailable (os error 11)`

This is typically because there are too many open file descriptors, and that's
possible given that `-j` doesn't limit the number of jobs that make will try to
run in parallel.

8 might still be too many, but let's try this value first, and then ramp down
until we find a consistently healthy CI sweet spot.