Swtich to pull_request_target for the trigger events
When targeting a `pull_request` event on a public repository, the permissions
are always restricted to read-only for github actions even if the `write`
permissions are explicity requested. This is to keep contributors from being
able to inject malicious code into the actions.

Technically, it should be fine to go back to the `REF` variable as this should
now be the ref for the target of the pull request which should be committed
code. But, it is also still fine to just use `refs/heads/master` in a hard-coded
fashion.