Add UnimplementedExecutionClient
In order to add flexibility to our testing approach, we need to control certain elements
of our environment.  One of these elements is the ExecutionClient.  The
ExecutionClient governs how information is recalled from, and flows through the
L2 in a predictable manner.

The ExecutionClient has a profound impact on our integration efforts as it is
directly responsible for message processing within the system.  In order to ensure
that our implementation makes sense, it is beneficial to be able to have a simplified
implementation of the ExecutionClient that behaves in a controllable / predicable
manner.

As a result, and to make it easier to implement some future implementations the
UnimplementedExecutionClient has been defined in order to make it easier to
write Mocks of the ExecutionClient without the need to fulfill the full
interface.