Add MessageHasher
In order to simplify Message processing, it is beneficial to have a consistent way to
generate a Hash for a given `MessageWithMetadata` input.  This way when we write
interactions / tests, we need not worry about the specifics of how we achieve these
hashes, provided that we do it consistently.

This commit adds a `MessageHasher` that does just this, and provides two separate
implementations for the convenience of testing.  It is likely that the user shouldn't
need to worry about which of these to use, so in addition a `DefaultMessageHasher`
has been defined so that the user doesn't need to worry about the specifics.

In general, and in the future, whenever a `MessageHasher` is needed, we will
default to the `DefaultMessageHasher`, while allowing the user to swap the
implementation out, should they need to.