Add MockEspressoKeyManager
The `EspressoKeyManagerInterface` plays an important role in the signing of
transaction data in order to authenticate and verify it when it comes back from
the Espresso Chain.  Creating the current implementation requires a minimum of
6 arguments, and it may be difficult to understand the how and why of each field
at a glance.  As a result, it is cumbersome to utilize and configure effectively.  Additionally, it provides not control over the private key that is utilized for signing
data, which makes testing with a result that is reproducible consistently difficult.

The `MockEspressoKeyManager` implements the `EspressoKeyManagerInterface`
interface, and allows some simple configuration options, with some sensible defaults
that can be overwritten by the user if desired.  This allows for a much easier to
setup `EspressoKeyManagerInterface`, that is only concerned with signing things
and does not worry about the other aspects.