Add MockEspressoChain
When testing our implementation that performs finalization checks against Espresso, it
is beneficial to operate on a controllable idealized version the Espresso Chain that
operates and documents our initial assumptions of how the chain should progress.
Additionally, it is beneficial to have direct control over this chain so that we are able
to control exactly when and how it advances, and the rate at which it produces blocks.

This simplified idealized representation of the `EspressoChain`, implements only the
methods needed for operation and utilization within the `TransactionStreamer`
interface.