Add EspressoClientSimulatedLatency
When interacting with an idealized `EspressoClient`, it becomes prudent to test
different scenarios in terms of network performance, and their affect on our
implementations progression.  As a result, it is beneficial to be able to simulate
latency that is incurred in the process via various means, such as Network
calls and the like.

The `EspressoClientSimulatedLatency` allows for the configuration of delays against
the three primarily utilized endpoints of the `EspressoClient` for the
`TransactionStreamer` interaction.  It also includes a delay that can be configured
against the Explorer endpoint, that is utilized as a temporary fix for performance issues
discovered live.