Add helper configuration functions for TransactionStreamer
Many of the fields within the `TransactionStreamer` that facilitate our Espresso
integration are a mix of public and private, which makes them difficult or impossible
to modify from an external package for testing purposes.

This file adds a simple way of configuring all of the configurable Espresso fields
without the need to make them specifically public.  It also helps to keep these
fields separate from the rest of the `TransactionStreamer`, so that the
`TransactionStreamer` specifics do not need to be exposed.