Add NewMockTransactionStreamerEnvironment function
In order to make the Espresso incorporation with the `TransactionStreamer` easier to
test, it helps to be able to easily spin up different scenarios for environments that
provides sensible defaults which allow basic invocation to work, while still allowing
for the flexibility to overwrite specifics.

The `NewMockTransactionStreamerEnvironment` provides the ability to do this.
Calling it without any options will still result in an implementation that can effectively
simulate these interactions, and allows the user to just focus on the results or the
specific criteria they are choosing to focus on, without needing to worry about
every individual detail.