Add TestTransactionStreamerEspressoThroughput
This test is setup to measure the implementation performance of the Espresso Chain
against the Transaction Streamer in the current integration.

This demonstrates how to use the mock setup to run the tests, and how we can
measure the performance using the utilities we have defined.  This also demonstrates
the configurable nature of the mock environment, and how one can utilize the
configuration to get meaningful measurements out of the system.