Add UntrustedSignerCert CustomDA test
This adds a test where an untrusted signer is used to sign the second
CustomDA batch that is posted to the SequencerInbox. The evil validator
node pretends that it is valid and processes it, whereas the good node
finds it invalid. Currently we don't have a way for the good node to
progress past an invalid certificate, so the test isn't able to continue,
but once we do then the test should trigger a dispute on the
ValidatePreimage proof for this certificate. We'll have to come back to
this test after that has been implemented.