Add a go workspace file
When working with a go project that has multiple go modules in it, you have to
tell the go tooling which modules are considered part of this workspace. The
other modules can still be used, but won't be considered part of the same
workspace when viewing them in VSCode or other tools that use gopls to reason
about a go project.