Refine Throughput Test to add test scenarios and throughput targets based on current implementation (#692)
* Add BlockBuilder for MockEspressoChain

In order to simulate real-world conditions it is beneficial to be able to simulate a
non-ideal block production that we currently see on the Espresso Chain.  The Blocks
that Espresso produces are essentially size constrained in order to ensure that the
production process doesn't tax the system more than it needs to be.

These restrictions can have an impact on the throughput rates we are able to
achieve.

As a result, it is beneficial to be able to simulate these restrictions in order to more
closely compare oujr tests against real-world scenarios.

* Revise TestTransactionStreamerEspressoThroughput conditions

The TestTransactionStreamerEspressoThroughput test originally was created just to
monitor throughput based on the parameters being configured for the mock
setups.  The idea was to determine how interval configuration impacted the integration
performance under idealized conditions.

With some base lines set, it is now important to try and ensure that our integration
continues to perform at an established baseline and well beyond in order to ensure
that we ware able to catch up in a sufficient amount of time, should we fall behind.

To achieve this goal the TestTransactionStreamerEspressoThroughput test has been
reworked to test specific scenarios against the required benchmarks.  Additionally
it is important to codify the requirements of it being able to outperform the baseline
by a factor.  In this case the factor chosen is around 8x, but this was chosen arbitrarily
and our tests have indicated that we are able to exceed this with the right polling
interval configuration.

* Remove redundant comments

Based on feedback provided by @sneh1999, comments that just state what is being
explicitly performed have been removed as they don't add anything meaningful to
what has been written / defined.

- https://github.com/EspressoSystems/nitro-espresso-integration/pull/692#discussion_r2247997061
- https://github.com/EspressoSystems/nitro-espresso-integration/pull/692#discussion_r2247999152

These comments are just stating what's already visibly being done.  They don't need
to exist in any meaningful way.   This commit removes them as per the suggestion.

* Rename Throughput test

In order to ensure that the Throughput is picked up by ci, it's beneficial to have its
name follow the existing `run-espresso-ci-job` scheme.  That is that it's name
should contain `TestEspresso`.  This alone won't ensure that it's evaluated in ci, but
it helps to be consistent in order to identify these tests more effectively.

* Modify tests being targeted by run-espresso-ci-job

The `run-espresso-ci-job` will only target tests that contain the name `TestEspresso`
in them, and are contained directly within the `system_tests` directory.  It would be
beneficial to target any `TestEspresso` tests that are a descendent of the `system_tests`
instead of just residing directly within it.

This change expands the tests being considered to be any descendents of `system_tests`
rather than just sitting in that single folder.