Add 10 minutes to bold tests (#3514)
For some reason, this is frequently timing out on CI builds.

The builds and tests are passing locally, so, I assume the reason the CI jobs
are failing is a timeout. The logs are far too verbose to understand if there is
some actual failure in CI.

Also, this commits some updates to the mod files generated by
`go get -v -t ./...` and updates the workflow not to pass the now-deprecated
`-d` option to `go get`