Refactor move multigascollector to arbos (#3524)
This allows docker buildx build nitro-node-dev . to build.
The reason is that that docker image builds the replay.wasm with strictly the directories that are supposed to contain code which affects the state transition function.
When the multigascollector package was in execution, it pulled in all of the dependencies of that go package.
So, things got a bit upside-down.