Change ValidatePreimage return type to u8 where possible
This changes the exported declaration of wavmio__validateCertificate to
have a u8 return type. In practice when it is built with the rust
wasm32-unknown-unknown target, sizes smaller than WASM's minimum size
type of i32 are extended. On the Go side, go:wasmimport only supports at
least 32 bit types in order to match the wasm ABI requirements.

So this change has no functional impact but it makes it slightly clearer
what the return type represents.