Add DA proof support to daprovider interface
This is part of a series of changes for the Custom DA project.

Summary:
- Added a new daprovider.Validator interface with methods for generating
  proofs for Custom DA systems.
- Added a reference implementation of a Custom DA provider.
- Added a daprovider factory that supports anytrust and referenceda
  modes. In a follow-up PR the nitro node startup sequence will be
  modified to use this. Currently only the separate daprovider server uses
  this.
- Replaced the AnyTrust-specific (aka das) provider server with a
  unified provider server that works with anytrust or referenceda modes.
- Extended the DA Client with new RPC methods for generating proofs.

Notes: The separate provider server executable is a thin RPC server
wrapper around the anytrust and referenceda implementations. The idea is
that people wanting to integrate their own DA system can use this as a
guide for how to implement their own RPC service that lives outside the
nitro codebase; we won't be including support for any additional DA
implementations in provider server executable that we distribute. For
legacy AnyTrust deployments we will most likely continue to have nitro
spawn the daprovider server in-process to avoid needing to run an extra
service, but by channeling everything through the JSON-RPC interface it
reduces surface area of what we have to support.

The Reference DA (referenceda) implementation is a minimal working
example of how one could implement a daprovider, including support for
validating the certificate against trusted signers. It uses an in-memory
data storage backend.

In various places there is commented out code related to the osp
contract bindings that haven't yet been committed to a nitro-contracts
branch that we want to use in nitro master.

This PR shouldn't change any existing functionality or behavior, except
the daprovider executable (which isn't used in production) has some
new configuration options:
```
 --mode               must be "anytrust" or "referenceda"
 --provider-server.*  server config eg address, port, etc
 --anytrust.*         was previosly called das-server
 --referenceda.*      referenceda specific options
```

As much as possible we will try to rename references to "das" to
"anytrust". When we launched Anytrust, we only had one offchain data
availability mode so we just called it "das" at the time.

This PR doesn't include new test code, but testing was done with the
end-to-end block validator and challenge system tests on the custom-da
branch.