Remove IsValidHeaderByte from RPC API
daprovider.Readers are now registered with their respective header bytes
at initialization. For external DA providers where the RPC client is used
this becomes single call at init to daprovider_getSupportedHeaderBytes.

Where previously in the inbox and block validator components we had
loops over each provider, checking each provider if they handle that
type of message, we now just check if we have a provider registered for
that message.

This means less RPC calls and makes the interface that providers need to
implement a bit simpler, with room for taking it out completely for
CustomDA if we can detect from other config that the external provider
is expected to be a CustomDA provider.