Overhaul GitHub actions and workflows to add a merge queue (#3655)
* Move the arbitrator tests into the central ci.yml

These tests only run when the files in certain directories have changed.

* Make yamllint happy

* Enable workflows on the main branch in addition to master

This is mostly for trying out the workflows along with realistic settings in
GitHub for a `main` branch.

* Fix ci workflows

* Fix the ci workflows again

* Fix where inputs goes in the action.yml

* The argument name is fetch-depth

* Build the deps first

* Fix Redis and Block Input runs

* Extract some reuable workflows

These can be called from higher-level workflows to reuse the code for pull
requset, merge_queue and nightly scheduled checks.

* Use ubuntu-4 instead of ubuntu-latest

The standard GitHub action runners are hitting up against the free limit too
often.

* Swap linting to after building

Otherwise, the contracts haven't been built yet and the linter can't find all
the generated sources it needs to build the go code.

* Require the build artifacts before go tests

This makes it possible to avoid rebuilding things that are already built.

* Don't use build artifacts at all

Making the long-running workflows wait on the fast-running build and lint job
only slows them both down and reduces parallelism.

* Fix the merge-check.yml for Design Approved

The previous version had a bug in how the after-next-release was being negated.

* Drop the branch filters from the merge group selector

The merge queue events actually happen on the merge branches which are never
going to be named `master` or `main`.

* Make Submodule Pin and Design Approved checks use sha

The previous workflow files wouldn't really work for merge_group triggered
workflow runs because the event type was different.

* Make Design Approved check only meaningful for PRs

The PR has labels, but the branch that the merge_queue runs on doesn't have
labels, and the event doesn't have a pull_request concept.

So, this change just makes the action succeed uncondiontally.

* Testing the main workflows (#3622)

This Pull Request is just to test that the workflows on the main branch are
still healthy.

I will probably merge it and some other harmless changes to the main branch, and
then open a PR to revert them all.

* Remove the main branch

For now, we're not going to try the switch from `master` to main as part of
adding the merge queue. It just makes the change too big. I'll get these changes
merged into the workflows on `master` and then I can enable the merge queue.

* Address review comments

* Checkout before calling ci-setup

Local actions cannot be used before the repository which holds them is checked
out.

* Switch back go golangci lint action and upgrade setup-go

There are two problems with "just" using the make lint command from CI.

  1. We need to take care of installing the golangci-lint binary.
	2. The GitHub action is much more efficiently integrated with the go caches.

We may want to split out the non-go linters into a separate make action at some
point soon, but it's not important right now.

* Add back a safety check for the branch

* Reenable these workflows for the main branch

This allows us to confirm that a merge queue set up on the main branch would
actually work as expected.

* Testing the merge queue config on main (#3667)

* Fix the merge workflow

This way, if the Arbitrator or Bold Legacy jobs are skipped (because the files
didn't change,) the can_proceed job will still succeed.

* This makes a minor change to a comment (#3668)

The reason is just to trigger the merge queue on the main branch merge.

* Address review feedback.

* Expand -t to --tag