Move funding common accounts to genesis block for L1
In the system tests, there is a particularly icky race condition which can cause
tests to fail if the L1Client and the L1Backend's TxPool objects end up with
different ideas about the balance of an account.

Essentially, when accounts like the "RollupOwner" account are funded just before
deploying the L1 contracts, and there is heavy CPU load on the machine running
the tests, it is possible to send the transaction for deploying the contract to
the backend which then attempts to validate it and sees that there is a 0
balance for the "RollupOwner" account, even though the L1.Client() sees the
account as funded.

Instead of funding the accounts with transactions, we just add the funds to them
when allocating the genesis block.

This is a better fix for: NIT-3886 because it continues to use the in-memory L1
node, but becomes deterministic for bootstrapping.