Enable ReferenceDA deployment for legacy rollup tests
Updated deployOnParentChain to support ReferenceDA with both BOLD and
legacy rollups. The Sequencer is configured as the trusted signer. This
fixes the TestBlockValidatorReferenceDA tests which validate block
validation with ReferenceDA mode for a legacy rollup

The legacy rollup contracts don't actually support a custom OSP set on
them, but this doesn't matter for testing validation only without a
challenge since the OSP contract is not used. The reference DA proof
validator contract is needed because it also is used by the DA reader to
validate certificates upon reading.