Removed the validateSeqMsg param from Reader
For DAS and ReferenceDA we almost always want to validate the
certificate. The only time we don't is in the replay binary when we want
to panic instead if it's invalid on the first read of the message,
otherwise we want to ignore it and assume it's already valid. So this
means that it doesn't make sense for validateSeqMsg to be part of the
reader API at all, which is great because it means we can simplify the
API. We can pass into the reader at construction time how to handle this
because in replay it always constructs a new (fake, non-network calling)
reader for each invocation.