Remove deprecated AnyTrust db storage service (#3701)
The DbStorageService has been deprecated for over a year, with a
warning printed for any remaining users during that time informing them
of how to migrate and that if they didn't, continuing to try to use it
would be a fatal error in future. The future is here.

This also lets us remove badgerdb is a dependency.