Add DA proof support to daprovider interface (#3600)
* Add DA proof support to daprovider interface

This is part of a series of changes for the Custom DA project.

Summary:
- Added a new daprovider.Validator interface with methods for generating
  proofs for Custom DA systems.
- Added a reference implementation of a Custom DA provider.
- Added a daprovider factory that supports anytrust and referenceda
  modes. In a follow-up PR the nitro node startup sequence will be
  modified to use this. Currently only the separate daprovider server uses
  this.
- Replaced the AnyTrust-specific (aka das) provider server with a
  unified provider server that works with anytrust or referenceda modes.
- Extended the DA Client with new RPC methods for generating proofs.

Notes: The separate provider server executable is a thin RPC server
wrapper around the anytrust and referenceda implementations. The idea is
that people wanting to integrate their own DA system can use this as a
guide for how to implement their own RPC service that lives outside the
nitro codebase; we won't be including support for any additional DA
implementations in provider server executable that we distribute. For
legacy AnyTrust deployments we will most likely continue to have nitro
spawn the daprovider server in-process to avoid needing to run an extra
service, but by channeling everything through the JSON-RPC interface it
reduces surface area of what we have to support.

The Reference DA (referenceda) implementation is a minimal working
example of how one could implement a daprovider, including support for
validating the certificate against trusted signers. It uses an in-memory
data storage backend.

In various places there is commented out code related to the osp
contract bindings that haven't yet been committed to a nitro-contracts
branch that we want to use in nitro master.

This PR shouldn't change any existing functionality or behavior, except
the daprovider executable (which isn't used in production) has some
new configuration options:
```
 --mode               must be "anytrust" or "referenceda"
 --provider-server.*  server config eg address, port, etc
 --anytrust.*         was previosly called das-server
 --referenceda.*      referenceda specific options
```

As much as possible we will try to rename references to "das" to
"anytrust". When we launched Anytrust, we only had one offchain data
availability mode so we just called it "das" at the time.

This PR doesn't include new test code, but testing was done with the
end-to-end block validator and challenge system tests on the custom-da
branch.

* Reduce code duplication in factory, more cert validation

* Fix for interface change on master

* Remove DACert flag from KnownHeaderBits for now

We will add DACertificateMessageHeaderFlag back to KnownHeaderBits in
a future PR; removing it for now avoids changing the replay binary
unnecessarily when merging in this PR.

* Remove preimageType from daprovider.Validator

* Remove IsValidHeaderByte from RPC API

daprovider.Readers are now registered with their respective header bytes
at initialization. For external DA providers where the RPC client is used
this becomes single call at init to daprovider_getSupportedHeaderBytes.

Where previously in the inbox and block validator components we had
loops over each provider, checking each provider if they handle that
type of message, we now just check if we have a provider registered for
that message.

This means less RPC calls and makes the interface that providers need to
implement a bit simpler, with room for taking it out completely for
CustomDA if we can detect from other config that the external provider
is expected to be a CustomDA provider.

* Move daprovider api types to own go pkg

* Use Promises with daprovider.Reader iface

* dasserver migration to provider_server

We moved the contents of dasserver to provider_server on the custom-da
branch and I accidentally left dasserver in place on this branch when
moving things across. I don't want to bring over the changes to the
daprovider construction logic in node.go yet so I just made the
das_migration.go file which we'll delete later.

* Split rpc payload and preimage methods

In normal execution we only care about the payload, and in validation we
only care about the preimages (we will reconstruct the payload using the
preimages). This change separates these concerns and simplifies the
signatures of these methods.

* Removed the validateSeqMsg param from Reader

For DAS and ReferenceDA we almost always want to validate the
certificate. The only time we don't is in the replay binary when we want
to panic instead if it's invalid on the first read of the message,
otherwise we want to ignore it and assume it's already valid. So this
means that it doesn't make sense for validateSeqMsg to be part of the
reader API at all, which is great because it means we can simplify the
API. We can pass into the reader at construction time how to handle this
because in replay it always constructs a new (fake, non-network calling)
reader for each invocation.

* Use Promises with Validator interface

* Add missing import

* Resgister DAS reader with both flag variants

* Use cancelable promises

* Fix reader registration for batch correctness checking

* Use promise for GetSupportedHeaderBytes too